+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-06-06 20:51:17
-// Commit: test(ui): add unit tests for ui business logic
+// Updated at: 2026-06-07 10:27:20
+// Commit: chore(ui): update dependencies and cleanup unused imports in ui
 
 class uiComponentState {
-  final String commitRef = '1818';
-  final String lastMessage = 'test(ui): add unit tests for ui business logic';
-  final DateTime timestamp = DateTime.parse('2026-06-06T20:51:17+05:00');
+  final String commitRef = '1829';
+  final String lastMessage = 'chore(ui): update dependencies and cleanup unused imports in ui';
+  final DateTime timestamp = DateTime.parse('2026-06-07T10:27:20+05:00');
 }